+++ docs/CHANGELOG.md
@@ -163,7 +163,7 @@ Version 1.2.0 (Willow): Not yet released
 - Fix Direct3D 11 renderer crashing when generating realtime bitmaps
 - Fix dynamic lights not rendering on meshes (fixed in Direct3D 11 only)
 - Fix dynamic lights not rendering on mover brushes (fixed in Direct3D 11 only)
-- Fix Direct3D 11 renderer not properly rendering vertex lighting for static meshes
+- Fix Direct3D 11 renderer not properly rendering vertex lighting for meshes
 - Fix `r_fullbright`, `r_lightmaps`, and `r_nearest` not working properly on Direct3D 11 renderer
 - Fix `Winner` and `Game Over` sounds played at the end of a match usually being wrong
 - Fix `Try Avoid Players` behaviour for spawn point selection not working properly